Includes block processing and storage capabilities for non-hybrid mode (pure PoW).
We have defined a process for executing a transaction to produces a series of transitions to be performed in other to change the world state and merkle tree. This also allows us to run mock state tree computation to produce and test the state root of a block without affecting the world state.

This is still work in progress. Future commits may significantly change things.