Implemented `ChainTransverser` and `WorldReader` to provide smart chain query capabilities.
`WorldReader` can search the best chain for an object if a specific chain is not known.

If a chain is know, it enables a miner mode that queries the chain and all ancestors of the chain, making sure that only objects known to the chain producer at the time are queried.