Added support for private RPC sessions. * Added `auth` and `login` RPC methods.
`auth` is an RPC method defined in the `rpc` package. It creates a session/auth token when a username and password matches the credentials in the config file.

`login` is a JS method defined in the `console` package. Internally, it calls `auth` and sets the token as the auth token to be used in the active console session.
* Introduction of `disableAuth` in rpc config to disable session token validation in RPC invocations. By default, this is set to false.

* Added more tests to `jsonrpc` package.